+++ aiohttp_sqlalchemy/views.py
@@ -4,7 +4,7 @@ from typing import Any, List, Optional
 from aiohttp.web import View
 from aiohttp_things.views import (
     ContextMixin,
-    InstanceMixin,
+    ItemMixin,
     ListMixin,
     PrimaryKeyMixin,
 )
@@ -64,9 +64,9 @@ class SAItemMixin(SAModelMixin, metaclass=ABCMeta):
     pass
 
 
-class SAItemAddMixin(SAItemMixin, InstanceMixin, metaclass=ABCMeta):
+class SAItemAddMixin(SAItemMixin, ItemMixin, metaclass=ABCMeta):
     def sa_add(self, *, key: Optional[str] = None) -> None:
-        self.sa_session(key).add(self.instance)
+        self.sa_session(key).add(self.item)
 
 
 class SAItemDeleteMixin(
@@ -83,7 +83,7 @@ class SAItemDeleteMixin(
 
 class SAItemEditMixin(
     SAItemMixin,
-    InstanceMixin,
+    ItemMixin,
     SAModelEditMixin,
     SAPrimaryKeyMixin,
     metaclass=ABCMeta,
@@ -96,7 +96,7 @@ class SAItemEditMixin(
 
 class SAItemViewMixin(
     SAItemMixin,
-    InstanceMixin,
+    ItemMixin,
     SAModelViewMixin,
     SAPrimaryKeyMixin,
     metaclass=ABCMeta,